feat/tournament-local (#17)
* teams now display & redirect to lan page from event page

* added to team when registered

* tournament list format

* list feature added

* README updated

* removed getChallongeParticipants

+++ app/EventTournamentParticipant.php
@@ -45,7 +45,8 @@ class EventTournamentParticipant extends Model
         self::created(function($model){
             if (
                 (!isset($model->event_tournament_team_id) || trim($model->event_tournament_team_id) == '') &&
-                (!$model->pug && $model->event_tournament_team_id == null)
+                (!$model->pug && $model->event_tournament_team_id == null) &&
+                $model->eventTournament->format != 'list'
             ) {
                 $challonge = new Challonge(env('CHALLONGE_API_KEY'));
                 $tournament = $challonge->getTournament($model->eventTournament->challonge_tournament_id);
@@ -55,18 +56,23 @@ class EventTournamentParticipant extends Model
                 }
                 $model->challonge_participant_id = $response->id;
                 $model->save();
-                return true;
             }
+            return true;
         });
         self::deleting(function($model){
-            if (!$model->pug && $model->event_tournament_team_id == null) {
+            if (!$model->pug && $model->event_tournament_team_id == null && $model->eventTournament->format != 'list') {
                 $challonge = new Challonge(env('CHALLONGE_API_KEY'));
                 $participant = $challonge->getParticipant($model->eventTournament->challonge_tournament_id, $model->challonge_participant_id);
                 if (!$response = $participant->delete()) {
                     return false;
                 }
-                return true;
             }
+            if ($model->tournamentTeam && $model->tournamentTeam->tournamentParticipants->count() == 1) {
+                if (!$model->tournamentTeam->delete()) {
+                    return false;
+                }
+            }
+            return true;
         });
     }
