+++ app/EventTournamentTeam.php
@@ -41,21 +41,25 @@ class EventTournamentTeam extends Model
     {
         parent::boot();
         self::created(function($model){
-            $challonge = new Challonge(env('CHALLONGE_API_KEY'));
-            $tournament = $challonge->getTournament($model->eventTournament->challonge_tournament_id);
-            if (!$response = $tournament->addParticipant(['participant[name]' => $model->name])) {
-                $model->delete();
-                return false;
+            if ($model->eventTournament->format != 'list') {
+                $challonge = new Challonge(env('CHALLONGE_API_KEY'));
+                $tournament = $challonge->getTournament($model->eventTournament->challonge_tournament_id);
+                if (!$response = $tournament->addParticipant(['participant[name]' => $model->name])) {
+                    $model->delete();
+                    return false;
+                }
+                $model->challonge_participant_id = $response->id;
+                $model->save();
             }
-            $model->challonge_participant_id = $response->id;
-            $model->save();
             return true;
         });
         self::deleting(function($model){
-            $challonge = new Challonge(env('CHALLONGE_API_KEY'));
-            $participant = $challonge->getParticipant($model->eventTournament->challonge_tournament_id, $model->challonge_participant_id);
-            if (!$response = $participant->delete()) {
-                return false;
+            if ($model->eventTournament->format != 'list') {
+                $challonge = new Challonge(env('CHALLONGE_API_KEY'));
+                $participant = $challonge->getParticipant($model->eventTournament->challonge_tournament_id, $model->challonge_participant_id);
+                if (!$response = $participant->delete()) {
+                    return false;
+                }
             }
             return true;
         });
